cmd/devp2p/internal/ethtest: move eth/72 tests to a new file
